Avoid a NPE when Jenkins contains folders
This is a very simple fix/workaround and renders folders inside the
Jenkins-panel with folder-icons. It doesn't add any functionality (such as
expanding/collapsing the folder), but it avoids the NPE caused by not finding an
icon for a folder.

NB: the fix could (should?) be more elegant, i.e. handling the concept of
folders instead of just performing a null-check and just giving back an icon